update tests for Agent
